Refactored + Improved Word-Scoring Logic
Playing the game previously revealed an edge case that produced errors,
where if a word, i.e. "GOGOS" is guessed for the word "ORGAN", both "G"
in "GOGOS" will be marked as a "partial" and a "correct", when there is
only one "G" in the wordle. The scoring logic was refactored to first
pass through only looking for correct letters, and uses a frequency
counter, which gets decremented with each letter occurence, to track
letters.